requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
discord.py==1.3.1
idna==2.9
multidict==4.7.5
pillow==7.0.0
pip==19.0.3
prettytable==0.7.2
python-dotenv==0.11.0
requests==2.23.0
setuptools==45.2.0
urllib3==1.25.8
websockets==8.1
yarl==1.4.2
